git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@3538 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/btree/btree.h
@@ -27,7 +27,7 @@ struct btree_node {
     uint64_t bytesize;
 };
 
-#define BTREE_MAGIC 0xcdef9425feed7980
+#define BTREE_MAGIC UINT64(0xcdef9425feed7980)
 
 struct btree {
     uint8_t order;
